Extract card offset calculation to a common place
